pyproject.toml
[project]
name = "py-wdp-001"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytest>=8.3.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
